impl Entry::version, and remove package submodule
